Address Bryan feedback
